scripts/automation/requirements.txt
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.3
